clean workspace created by job (#1032)
* clean workspace created by job

* Update summary_test.py

* Update summary_test.py

* update dependencies for tensorboard

* update dependencies for tensorboard

* specify precision for int type

+++ benchmark/fluid/ES/utils.py
@@ -65,7 +65,7 @@ def unflatten(flat_array, array_shapes):
     i = 0
     arrays = []
     for shape in array_shapes:
-        size = np.prod(shape, dtype=np.int)
+        size = np.prod(shape, dtype=np.int32)
         array = flat_array[i:(i + size)].reshape(shape)
         arrays.append(array)
         i += size
