+++ benchmark/torch/AlphaZero/submission_template.py
@@ -202,7 +202,7 @@ class Board():
         self.win_length = win_length or DEFAULT_WIN_LENGTH
 
         if np_pieces is None:
-            self.np_pieces = np.zeros([self.height, self.width], dtype=np.int)
+            self.np_pieces = np.zeros([self.height, self.width], dtype=np.int32)
         else:
             self.np_pieces = np_pieces
             assert self.np_pieces.shape == (self.height, self.width)
